issue dependancies working

+++ reimp/src/components/IssueDependancies.js
@@ -9,7 +9,10 @@ import {
     removeIssueNeedsAnother
 } from '../actions/Issues'
 import IssueName from './IssueName'
+import IssueStatus from './IssueStatus'
 import IssueSelectorForm from './form/IssueSelectorForm'
+import ModalDialog from './ModalDialog'
+import PopupPanelHeading from './PopupPanelHeading'
 
 class IssueDependancies extends Component {
 
@@ -47,30 +50,36 @@ class IssueDependancies extends Component {
                        adding_issue_needs_us: false})
     }
 
-    onCreateIssueNeedingUs(new_values) {
+    onCreateNeedsIssue(new_values) {
         const { dispatch, issue_id } = this.props
         const other_issue_id = new_values.issue_id
         dispatch(addIssueNeedsAnother(issue_id, other_issue_id))
-        this.onStopSelectingIssue()
+        this.onStopAddingIssueDependancy()
     }
     
-    onCreateNeedsIssue(new_values) {
+    onCreateIssueNeedingUs(new_values) {
         const { dispatch, issue_id } = this.props
         const other_issue_id = new_values.issue_id
         dispatch(addIssueNeedsAnother(other_issue_id, issue_id))
         this.onStopAddingIssueDependancy()
     }
     
-    onDeleteIssueNeedingUs(other_issue_id) {
-        const { dispatch, issue_id } = this.props
-        dispatch(removeIssueNeedsAnother(issue_id, other_issue_id))
-    }
-
     onDeleteNeedsIssue(other_issue_id) {
         const { dispatch, issue_id } = this.props
-        dispatch(removeIssueNeedsAnother(other_issue_id, issue_id))
+        if ( ! window.confirm("Remove this dependancy?" ) ) {
+            return false
+        }
+        dispatch(removeIssueNeedsAnother(issue_id, other_issue_id))
     }
     
+    onDeleteIssueNeedingUs(other_issue_id) {
+        const { dispatch, issue_id } = this.props
+        if ( ! window.confirm("Remove this dependancy?" ) ) {
+            return false
+        }
+        dispatch(removeIssueNeedsAnother(other_issue_id, issue_id))
+    }
+
     refresh(props) {
         const {dispatch, issue_id, issue} = props
         dispatch(ensureIssuesLoaded([issue_id]))
@@ -86,38 +95,58 @@ class IssueDependancies extends Component {
         const { issue } = this.props
         const { adding_issue_needs_us, adding_needs_issue } = this.state
         return (
-            <PropertyStackComponent title="Dependancies">
+            <PropertyStackComponent>
               <div className="property-label">
                 Other issues needed by this issue
               </div>
               <div className="property-value">
-                { map(issue.needs_issues, (issue_id) =>
+                { map(issue.needs_issue_ids, (issue_id) =>
                     <div className="property-row">
                       <IssueName issue_id={issue_id} />
+                      <IssueStatus issue_id={issue_id} />
                       <div onClick={() => this.onDeleteNeedsIssue(issue_id)}
                            className="icon--small-delete" />
                     </div>
                   )}
-                    <button className="button button--primary" onClick={this.onStartAddingIssueDependancy}>Add</button>
+                    <div className="icon--add" onClick={this.onStartAddingNeedsIssue}></div>
               </div>
               <div className="property-label">
                 Other issues that need this issue
               </div>
               <div className="property-value">
-                { map(issue.issues_needing_us, (issue_id) =>
+                { map(issue.issue_ids_needing_us, (issue_id) =>
                     <div className="property-row">
                       <IssueName issue_id={issue_id} />
+                      <IssueStatus issue_id={issue_id} />
                       <div onClick={() => this.onDeleteIssueNeedingUs(issue_id)}
                            className="icon--small-delete" />
                     </div>
                   )}
               </div>
-              <button className="button button--primary" onClick={this.onStartAddingIssueDependancy}>Add</button>
-              { adding_issue_needs_us &&
-                <IssueSelectorForm onSubmitted={this.onCreateIssueNeedingUs} />
-              }
+              <div className="icon--add" onClick={this.onStartAddingIssueNeedingUs}></div>
               { adding_needs_issue &&
-                <IssueSelectorForm onSubmitted={this.onCreateNeedsIssue} />
+                <ModalDialog isOpen={true}
+                             onClose={this.onStopAddingIssueDependancy}
+                             title={"Adding issue dependancy"}
+                             variant="large">
+                  <PopupPanelHeading>
+                    Select an issue that is needed by issue
+                    <IssueName issue_id={issue.id} />
+                  </PopupPanelHeading>
+                  <IssueSelectorForm onSubmitted={this.onCreateNeedsIssue} />
+                </ModalDialog>
+              }
+              { adding_issue_needs_us &&
+                <ModalDialog isOpen={true}
+                             onClose={this.onStopAddingIssueDependancy}
+                             title={"Adding issue dependancy"}
+                             variant="large">
+                  <PopupPanelHeading>
+                    Select an issue that needs issue
+                    <IssueName issue_id={issue.id} />
+                  </PopupPanelHeading>
+                  <IssueSelectorForm onSubmitted={this.onCreateIssueNeedingUs} />
+                </ModalDialog>
               }
             </PropertyStackComponent>
         )
